Keep component unopinionated
