+++ lib/components/component.dart
@@ -1,12 +1,7 @@
 import 'dart:ui';
 
 import 'package:flutter/painting.dart';
-import 'package:meta/meta.dart';
-import 'package:ordered_set/comparing.dart';
-import 'package:ordered_set/ordered_set.dart';
 
-import '../effects/effects.dart';
-import '../effects/effects_handler.dart';
 import '../extensions/vector2.dart';
 import '../game.dart';
 
@@ -14,13 +9,8 @@ import '../game.dart';
 ///
 /// Components can be bullets flying on the screen, a spaceship or your player's fighter.
 /// Anything that either renders or updates can be added to the list on [BaseGame]. It will deal with calling those methods for you.
-/// Components also have other methods that can help you out if you want to overwrite them.
+/// Components also have other methods that can help you out if you want to override them.
 abstract class Component {
-  final EffectsHandler _effectsHandler = EffectsHandler();
-
-  final OrderedSet<Component> children =
-      OrderedSet(Comparing.on((c) => c.priority));
-
   /// Whether this component has been loaded yet. If not loaded, [BaseGame] will not try to render it.
   ///
   /// Sprite based components can use this to let [BaseGame] know not to try to render when the [Sprite] has not been loaded yet.
@@ -52,13 +42,10 @@ abstract class Component {
   /// The time [t] in seconds (with microseconds precision provided by Flutter) since the last update cycle.
   /// This time can vary according to hardware capacity, so make sure to update your state considering this.
   /// All components on [BaseGame] are always updated by the same amount. The time each one takes to update adds up to the next update cycle.
-  @mustCallSuper
-  void update(double dt) {
-    _effectsHandler.update(dt);
-  }
+  void update(double dt) {}
 
   /// Renders this component on the provided Canvas [c].
-  void render(Canvas c);
+  void render(Canvas c) {}
 
   /// It receives the new game size.
   /// Executed right after the component is attached to a game and right before [onMount] is called
@@ -77,64 +64,4 @@ abstract class Component {
 
   /// Called right before the component is removed from the game
   void onRemove() {}
-
-  /// Called to check whether the point should be counted as a tap on the component
-  bool checkOverlap(Vector2 point) => false;
-
-  /// Add an effect to the component
-  void addEffect(ComponentEffect effect) {
-    _effectsHandler.add(effect, this);
-  }
-
-  /// Mark an effect for removal on the component
-  void removeEffect(ComponentEffect effect) {
-    _effectsHandler.removeEffect(effect);
-  }
-
-  /// Remove all effects
-  void clearEffects() {
-    _effectsHandler.clearEffects();
-  }
-
-  /// Get a list of non removed effects
-  List<ComponentEffect> get effects => _effectsHandler.effects;
-
-  /// Uses the game passed in to prepare the child component before it is added
-  /// to the list of children
-  void addChild(Game gameRef, Component c) {
-    if (gameRef is BaseGame) {
-      gameRef.prepare(c);
-    }
-    children.add(c);
-  }
-
-  /// This method first calls the passed function on itself and then recursively propagates it to every children
-  /// and grandchildren (and so on) of this component, either until it has propagated through the whole tree or
-  /// if the handler at any point returns false.
-  ///
-  /// This method is important to be used by the engine to propagate actions like rendering, taps, etc,
-  /// but you can call it yourself if you need to apply an action to the whole component chain.
-  /// It will only consider components of type T in the hierarchy, so use T = Component to target everything.
-  bool propagateToChildren<T extends Component>(
-    bool Function(T) handler,
-  ) {
-    bool shouldContinue = true;
-    if (this is T) {
-      shouldContinue = handler(this as T);
-      if (!shouldContinue) {
-        return false;
-      }
-    }
-    for (Component child in children) {
-      if (child is T) {
-        shouldContinue = handler(child);
-        if (shouldContinue) {
-          shouldContinue = child.propagateToChildren(handler);
-        } else {
-          break;
-        }
-      }
-    }
-    return shouldContinue;
-  }
 }
